requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==3.7.1
attrs==23.1.0
audioread==3.0.1
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
dataclasses-json==0.6.2
decorator==5.1.1
docstring-parser==0.15
Flask==3.0.0
frozenlist==1.4.0
grpcio==1.54.0
idna==3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonref==1.1.0
langchain==0.0.340
langsmith==0.0.66
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
packaging==23.2
pathlib==1.0.1
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
protobuf==3.20.3
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.17.2
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
scikit-learn==1.3.2
scipy==1.11.4
sievedata==0.1.0.0.post8
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.23
sseclient==0.0.27
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.64.1
typeguard==4.1.5
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.3
youtube-transcript-api==0.6.1
yt-dlp==2023.11.16
